refactor(api-reference): fix persistence-plugin filename typo
